feat(web): central write-gate middleware + client scope stamp (G4)
Read-only enforcement layer 2 for the showcase demo project (spec task G4).
Adds one global `functionMiddleware` entry in `start.ts` whose two halves —
`.client()` stamps the current `ProjectScope` onto every outgoing server-fn
request; `.server()` rejects writes under a read-only scope with a typed
`ReadOnlyProjectError`.

Method discipline makes the gate reliable: all GETs pass, POSTs are treated as
writes minus a 3-item POST-read allowlist (previewEvaluation,
listLinearTeamsForApiKey, listCursorRepositoriesForApiKey). Fails closed on an
unidentifiable POST.

Behavior-neutral: the gate only fires under showcase/public scope, and nothing
sets a non-`live` scope yet. `ProjectScope` is a minimal live-default source
(coordination seam for task G3, which will own the full scope context);
`ReadOnlyProjectError` lives in `@domain/shared` so it is client-safe for the
G1 mutation-error handler to branch on.

Umbrella PR: #3821


WIP

WIP

WIP

WIP
