+++ apps/web/src/domains/projects/project-scope.tsx
@@ -3,46 +3,31 @@ import { createContext, type ReactNode, use } from "react"
 /**
  * Ambient "which project surface am I reading data for?" scope.
  *
- * Generalizes the original sandbox-only trace scope so the same reused read
- * collections (traces, sessions, tools, spans) serve Live, the Test-Mode
- * sandbox, and — once wired — the shared read-only Showcase. Each non-live
- * scope gets its own query-key namespace (so one scope's cache never bleeds
- * into another's) and, where the server needs it, its own server-fn payload
- * fragment. Components calling the collection hooks don't change; the scope is
- * read from context.
+ * Generalizes the sandbox-only trace scope so the same reused read collections
+ * (traces, sessions, tools, spans) serve Live, the Test-Mode sandbox, and the
+ * shared read-only Showcase. Each non-live scope gets its own query-key
+ * namespace (so one scope's cache never bleeds into another's) and, where the
+ * server needs it, its own server-fn payload fragment. Components calling the
+ * collection hooks don't change; the scope is read from context.
  *
  * `live` is the default and a strict no-op: an empty key prefix and an empty
  * payload fragment, so every query key and request under Live stays
- * BYTE-IDENTICAL to production before this scope existed. Only non-live kinds
- * add a namespace.
- *
- * Ownership: G3 owns this primitive. G4 consumes it (client scope stamp +
- * read-only write-gate) and S3 sets `showcase` from the reserved-slug loader.
- * `sandbox` is the only non-live kind wired today; `showcase` is declared here
- * for those tasks and is otherwise dormant.
+ * byte-identical to an unscoped read. Only non-live kinds add a namespace.
  *
  * `sandbox` carries a client-provided org id (authorized server-side by the
  * sandbox middleware). `showcase` carries no org id — its org is resolved
  * server-side from the showcase pointer, never trusted from the client.
- *
- * @public Consumed by G4/S3; not imported elsewhere until then.
  */
 export type ProjectScope =
   | { readonly kind: "live" }
   | { readonly kind: "sandbox"; readonly orgId: string }
   | { readonly kind: "showcase" }
 
-/**
- * The default scope. A stable reference so Live consumers keep memo identity.
- * @public Consumed by G4/S3; not imported elsewhere until then.
- */
+/** The default scope. A stable reference so Live consumers keep memo identity. */
 export const LIVE_SCOPE: ProjectScope = { kind: "live" }
 
-/**
- * Read with {@link useProjectScope} (defaults to {@link LIVE_SCOPE}).
- * @public G4's client scope stamp reads the raw context; not imported elsewhere until then.
- */
-export const ProjectScopeContext = createContext<ProjectScope>(LIVE_SCOPE)
+/** The scope context. Read it with {@link useProjectScope} (defaults to {@link LIVE_SCOPE}). */
+const ProjectScopeContext = createContext<ProjectScope>(LIVE_SCOPE)
 
 export function ProjectScopeProvider({ scope, children }: { scope: ProjectScope; children: ReactNode }) {
   return <ProjectScopeContext.Provider value={scope}>{children}</ProjectScopeContext.Provider>
@@ -87,4 +72,22 @@ export function projectScopeData(scope: ProjectScope): { sandboxOrgId?: string }
  */
 export function sandboxOrgIdForScope(scope: ProjectScope): string | undefined {
   return scope.kind === "sandbox" ? scope.orgId : undefined
+}
+
+/**
+ * Whether the scope forbids writes. Only `showcase` (the shared read-only demo)
+ * is read-only; `sandbox` stays writable.
+ */
+export function isReadOnlyScope(scope: ProjectScope): boolean {
+  return scope.kind === "showcase"
+}
+
+/**
+ * The current scope for consumers that run outside React and so cannot read it
+ * via {@link useProjectScope} — notably the client-side write-gate middleware,
+ * which stamps the scope onto outgoing server-fn requests. Defaults to
+ * {@link LIVE_SCOPE}.
+ */
+export function getCurrentProjectScope(): ProjectScope {
+  return LIVE_SCOPE
 }
