+++ knip.json
@@ -167,10 +167,6 @@
   "ignoreBinaries": [
     "tmuxinator",
     "pulumi",
-    "tsx",
-    "tsdown",
-    "mcp-inspector",
-    "email",
     "packages/cli/target/release/latitude"
   ],
   "ignoreDependencies": [
@@ -180,11 +176,7 @@
     "@pulumi/datadog",
     "@pulumi/random",
     "fern-api",
-    "chdb",
-    "tsx",
-    "zod",
-    "@modelcontextprotocol/inspector",
-    "react-email"
+    "chdb"
   ],
   "tags": [
     "-knipignore"
